+++ cart/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from django.contrib.auth.models import User
+
 
 from shop.models import Product, Service
 
@@ -29,15 +29,3 @@ class CartItem(models.Model):
 
 	def __str__(self):
 		return self.product.name
-	
-	
-class UserProfile(models.Model):
-	'''
-	A customer profile
-	'''
-	user = models.OneToOneField(User, on_delete=models.CASCADE)
-	address = models.CharField(max_length=250, blank=True)
-	city = models.CharField(max_length=250, blank=True)
-	postcode = models.CharField(max_length=10, blank=True)
-	country = models.CharField(max_length=200, blank=True)
-
