+++ cart/templates/profile.html
@@ -9,9 +9,6 @@
 	Update shipping address - Perfect Cushion Store
 {% endblock %}
 {% block content %}
-	<form method="POST">
-      {% csrf_token %}
-      {{ form|crispy }}
-      <button type="submit">Update</button>
-    </form>
+	{% crispy form  %}
+	<div id='form-errors'>{{ form.errors }}</div>
 {% endblock %}
