+++ order/models.py
@@ -1,4 +1,6 @@
 from django.db import models
+from django.contrib.auth.models import User
+
 
 class Order(models.Model):
 	token = models.CharField(max_length=250, blank=True)
@@ -28,3 +30,17 @@ class OrderItem(models.Model):
 
 	def __str__(self):
 		return self.product
+	
+	
+	
+class UserProfile(models.Model):
+	'''
+	A customer profile
+	'''
+	user = models.OneToOneField(User, on_delete=models.CASCADE)
+	address = models.CharField(max_length=250, blank=True)
+	city = models.CharField(max_length=250, blank=True)
+	postcode = models.CharField(max_length=10, blank=True)
+	country = models.CharField(max_length=200, blank=True)
+	
+
